perf: extract recharts into dynamic import wrappers
Bundle recharts behind next/dynamic boundaries to prevent its
large module graph from being included in the initial JS payload.

- CostOverviewTab.tsx → dynamic(() => import('./components/CostCharts'))
- ProviderUtilizationTab.tsx → dynamic(() => import('./components/ProviderCharts'))
- BurnRateChart.tsx → dynamic(() => import('./components/BurnRateChartInner'))
- Created 3 wrapper files with 'use client' and all recharts imports

Reduces initial bundle by ~35 kB (recharts + dependencies).
