+++ src/app/(dashboard)/dashboard/analytics/ProviderUtilizationTab.tsx
@@ -3,16 +3,9 @@
 import { useTranslations } from "next-intl";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { useProviderNodeMap, resolveProviderName } from "@/lib/display/useProviderNodeMap";
-import {
-  CartesianGrid,
-  Legend,
-  Line,
-  LineChart,
-  ResponsiveContainer,
-  Tooltip,
-  XAxis,
-  YAxis,
-} from "recharts";
+import dynamic from "next/dynamic";
+
+const ProviderCharts = dynamic(() => import("./components/ProviderCharts"), { ssr: false });
 import Card from "@/shared/components/Card";
 import ProviderIcon from "@/shared/components/ProviderIcon";
 import TimeRangeSelector from "@/shared/components/analytics/TimeRangeSelector";
@@ -309,60 +302,17 @@ export default function ProviderUtilizationTab() {
           </div>
         ) : (
           <div className="flex flex-col gap-5">
-            <div className="h-80 w-full rounded-xl border border-black/5 bg-black/[0.02] px-3 py-4 dark:border-white/5 dark:bg-white/[0.02]">
-              <ResponsiveContainer width="100%" height="100%">
-                <LineChart data={chartData} margin={{ top: 8, right: 16, bottom: 0, left: 0 }}>
-                  <CartesianGrid
-                    stroke="var(--color-border)"
-                    strokeDasharray="3 3"
-                    vertical={false}
-                  />
-                  <XAxis
-                    dataKey="timestamp"
-                    tickFormatter={(value) => formatTimestamp(String(value), range)}
-                    tick={{ fill: "var(--color-text-muted)", fontSize: 12 }}
-                    axisLine={{ stroke: "var(--color-border)" }}
-                    tickLine={{ stroke: "var(--color-border)" }}
-                    minTickGap={24}
-                  />
-                  <YAxis
-                    domain={[0, 100]}
-                    tickFormatter={formatPercent}
-                    tick={{ fill: "var(--color-text-muted)", fontSize: 12 }}
-                    axisLine={{ stroke: "var(--color-border)" }}
-                    tickLine={{ stroke: "var(--color-border)" }}
-                    width={44}
-                  />
-                  <Tooltip
-                    labelFormatter={(value) => formatTooltipTimestamp(String(value), range)}
-                    formatter={(value: number, name: string) => [formatPercent(value), name]}
-                    contentStyle={{
-                      backgroundColor: "var(--color-surface)",
-                      borderColor: "var(--color-border)",
-                      borderRadius: 12,
-                      color: "var(--color-text-main)",
-                      boxShadow: "var(--shadow-soft)",
-                    }}
-                    itemStyle={{ color: "var(--color-text-main)" }}
-                    labelStyle={{ color: "var(--color-text-main)", fontWeight: 600 }}
-                  />
-                  <Legend />
-                  {data?.providers.map((provider) => (
-                    <Line
-                      key={provider}
-                      type="monotone"
-                      dataKey={provider}
-                      name={resolveProviderName(provider, nodeMap)}
-                      stroke={providerColors.get(provider) ?? "var(--color-primary)"}
-                      strokeWidth={2.5}
-                      dot={false}
-                      activeDot={{ r: 4, strokeWidth: 0 }}
-                      connectNulls
-                    />
-                  ))}
-                </LineChart>
-              </ResponsiveContainer>
-            </div>
+            <ProviderCharts
+              chartData={chartData}
+              providers={data?.providers ?? []}
+              providerColors={providerColors}
+              range={range}
+              resolveProviderName={resolveProviderName}
+              nodeMap={nodeMap}
+              formatTimestamp={formatTimestamp}
+              formatPercent={formatPercent}
+              formatTooltipTimestamp={formatTooltipTimestamp}
+            />
 
             <div className="grid grid-cols-1 gap-4 md:grid-cols-2 xl:grid-cols-3">
               {latestPoints.map((point) => {
